Moving domain objects
Making them more distinct by moving them to a folder and updating their
namespace.
