--- lib/D2L.Extensibility.AuthSdk.IntegrationTests/Organization.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace D2L.Extensibility.AuthSdk.IntegrationTests {
-	[Serializable]
-	public sealed class Organization {
-		public long Identifier { get; set; }
-		public string Name { get; set; }
-	}
-}
