--- lib/D2L.Extensibility.AuthSdk.IntegrationTests/ProductVersions.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace D2L.Extensibility.AuthSdk.IntegrationTests {
-	[Serializable]
-	public sealed class ProductVersions {
-		public string LatestVersion { get; set; }
-		public string ProductCode { get; set; }
-		public string[] SupportedVersions { get; set; }
-	}
-}
